Rename ItemReducer to itemReducer
